fix: int->double cast for _dailyGoal, remove unused imports in profile_screen

+++ lib/main.dart
@@ -261,7 +261,7 @@ class _MyHomePageState extends State<MyHomePage> with TickerProviderStateMixin {
     // Обновляем виджет на домашнем экране
     await HomeWidgetService().update(
       currentMl: _todayWater,
-      goalMl: _dailyGoal,
+      goalMl: _dailyGoal.toDouble(),
     );
 
     // Пробуем сразу отправить на сервер
